# specs/001-social-media-marketing/contracts/api-spec.yaml
openapi: 3.0.3
info:
  title: Social Media Marketing Dashboard API
  description: API for managing product briefs, assets, creative ideas, and social media content generation
  version: 1.0.0
servers:
  - url: http://localhost:8002
    description: Local development server
paths:
...
  /ideas/{idea_id}:
    get:
      summary: Get a specific idea
      operationId: getIdea
      parameters:
        - $ref: '#/components/parameters/IdeaId'
      responses:
        '200':
          description: Idea details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Idea'
        '404':
          $ref: '#/components/responses/NotFound'
  /ideas/{idea_id}/regenerate:
    post:
      summary: Regenerate a new idea to replace current one
      operationId: regenerateIdea
      parameters:
        - $ref: '#/components/parameters/IdeaId'
      responses:
        '200':
          description: New idea generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Idea'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
  /ideas/{idea_id}/generate-creative:
    post:
      summary: Generate final creative using Adobe Firefly
      operationId: generateCreative
      parameters:
        - $ref: '#/components/parameters/IdeaId'
      responses:
        '201':
          description: Creative generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Creative'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
...
components:
  parameters:
...
    IdeaId:
      name: idea_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    Idea:
      type: object
      required:
        - id
        - brief_id
        - region
        - demographic
        - content
        - language_code
        - generation_count
        - created_at
      properties:
        id:
          type: string
          format: uuid
        brief_id:
          type: string
          format: uuid
        region:
          type: string
        demographic:
          type: string
        content:
          type: string
        language_code:
          type: string
        generation_count:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Creative:
      type: object
      required:
        - id
        - idea_id
        - file_path
        - mime_type
        - file_size
        - generation_count
        - created_at
      properties:
        id:
          type: string
          format: uuid
        idea_id:
          type: string
          format: uuid
        file_path:
          type: string
        mime_type:
          type: string
        file_size:
          type: integer
        firefly_job_id:
          type: string
          nullable: true
        generation_count:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Error:
      type: object
      required:
        - message
      properties:
        message:
          type: string
        detail:
          type: string
          nullable: true
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ServerError:
      description: Server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'